issues with delete and upadate urls

+++ coreApp/views.py
@@ -31,14 +31,17 @@ def property_model_list(request):
         "listings":listings
     }
     return render(request, "listings_list.html", context)
-#retriving the properties.
 
+
+#retrieving the properties.
 def listing_retrieve(request, pk):
     listing = Property_model.objects.get(id=pk)
     context = {
         "listing": listing
     }
     return render (request,"listing_retrieve.html", context)
+
+
 
 def listing_create(request):
     form = ListingForm()
@@ -61,12 +64,15 @@ def listing_update(request, pk):
         if form.is_valid():
             form.save()
             return redirect("/")
-    context = {
-        "form": form
-    }
+        
+    context = {"form": form}
+
     return render(request, "listing_update.html", context)
 
 def listing_delete(request, pk):
     listing = Property_model.objects.get(id=pk)
-    listing.delete()
-    return redirect("/")
+    if request.method == "POST":
+        listing.delete()
+        return redirect("/")
+    context ={'listing':listing}
+    return render (request, 'delete.html', context)
